Restructure audit module
