+++ src/client/app/admin/index.ts
@@ -1,6 +1,7 @@
 export * from './admin.component';
 
 export * from './admin.service';
+export * from './users.service';
 
 export * from './admin.module';
 export * from './admin-routing.module';
