+++ src/client/app/audit/audit.service.ts
@@ -1,8 +1,8 @@
 import {Injectable} from '@angular/core';
 import {Response} from '@angular/http';
 import {Observable} from 'rxjs/Observable';
-import { AsyHttp, HttpOptions } from 'app/shared/asy-http.service';
-import { PagingOptions } from 'app/shared/pager.component';
+
+import { AsyHttp, HttpOptions, PagingOptions } from 'app/shared';
 
 @Injectable()
 export class AuditService {
